+++ src/main/java/me/diffusehyperion/Client/ClientServerSFTPDetails.java
@@ -4,18 +4,18 @@ import org.json.simple.JSONObject;
 
 public class ClientServerSFTPDetails {
     private final String ip;
-    private final Long port;
+    private final int port;
 
     public ClientServerSFTPDetails(JSONObject object) {
         ip = (String) object.get("ip");
-        port = (Long) object.get("port");
+        port = ((Long) object.get("port")).intValue();
     }
 
     public String getIp() {
         return ip;
     }
 
-    public Long getPort() {
+    public int getPort() {
         return port;
     }
 }
